move S3 collector into the collectors package
